% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{write_cf_outputs}
\alias{write_cf_outputs}
\title{Write decomposition outputs}
\usage{
write_cf_outputs(object, output_dir)
}
\arguments{
\item{object}{A `cfdecomp` object.}

\item{output_dir}{Destination directory.}
}
\value{
Invisibly returns `output_dir`.
}
\description{
Write decomposition outputs
}
